# Conflicts:
#	products/ASC.Files/Core/Services/DocumentService/Configuration.cs
#	products/ASC.Files/Core/Services/DocumentService/DocumentServiceTracker.cs
#	products/ASC.Files/Core/Services/WCFService/FileOperations/FileDownloadOperation.cs

+++ products/ASC.Files/Core/Services/DocumentService/DocumentServiceConnector.cs
@@ -1,4 +1,4 @@
-// (c) Copyright Ascensio System SIA 2009-2024
+// (c) Copyright Ascensio System SIA 2009-2025
 // 
 // This program is a free software product.
 // You can redistribute it and/or modify it under the terms
@@ -61,7 +61,6 @@ public class DocumentServiceConnector(ILogger<DocumentServiceConnector> logger,
         try
         {
             return await ASC.Files.Core.Helpers.DocumentService.GetConvertedUriAsync(
-                fileUtility,
                 filesLinkUtility.DocServiceConverterUrl,
                 documentUri,
                 fromExtension,
@@ -73,7 +72,9 @@ public class DocumentServiceConnector(ILogger<DocumentServiceConnector> logger,
                 spreadsheetLayout,
                 options,
                 isAsync,
-                fileUtility.SignatureSecret,
+                filesLinkUtility.DocServiceSignatureSecret,
+                filesLinkUtility.DocServiceSignatureHeader,
+                await filesLinkUtility.GetDocServiceSslVerificationAsync(),
                 clientFactory,
                 toForm);
         }
@@ -90,19 +91,20 @@ public class DocumentServiceConnector(ILogger<DocumentServiceConnector> logger,
                                string[] users = null,
                                MetaData meta = null)
     {
-        logger.DebugDocServiceCommand(method.ToStringFast(), fileId.ToString(), docKeyForTrack, callbackUrl, users != null ? string.Join(", ", users) : "null", JsonConvert.SerializeObject(meta));
+        logger.DebugDocServiceCommand(method.ToStringFast(), fileId.ToString(), docKeyForTrack, callbackUrl, users != null ? string.Join(", ", users) : "null", JsonSerializer.Serialize(meta));
         
         try
         {
             var commandResponse = await CommandRequestAsync(
-                fileUtility,
                 filesLinkUtility.DocServiceCommandUrl,
                 method,
                 GenerateRevisionId(docKeyForTrack),
                 callbackUrl,
                 users,
                 meta,
-                fileUtility.SignatureSecret,
+                filesLinkUtility.DocServiceSignatureSecret,
+                filesLinkUtility.DocServiceSignatureHeader,
+                await filesLinkUtility.GetDocServiceSslVerificationAsync(),
                 clientFactory);
 
             if (commandResponse.Error == ErrorTypes.NoError)
@@ -127,15 +129,25 @@ public class DocumentServiceConnector(ILogger<DocumentServiceConnector> logger,
         string scriptUrl = null;
         if (!string.IsNullOrEmpty(inputScript))
         {
-            using (var stream = new MemoryStream())
-            await using (var writer = new StreamWriter(stream))
+            if (System.IO.File.Exists(inputScript))
             {
-                await writer.WriteAsync(inputScript);
-                await writer.FlushAsync();
-                stream.Position = 0;
-                scriptUrl = await pathProvider.GetTempUrlAsync(stream, ".docbuilder");
+                await using (var stream = System.IO.File.OpenRead(inputScript))
+                {
+                    scriptUrl = await pathProvider.GetTempUrlAsync(stream, ".docbuilder");
+                }
             }
-            scriptUrl = await ReplaceCommunityAddressAsync(scriptUrl);
+            else
+            {
+                using (var stream = new MemoryStream())
+                await using (var writer = new StreamWriter(stream))
+                {
+                    await writer.WriteAsync(inputScript);
+                    await writer.FlushAsync();
+                    stream.Position = 0;
+                    scriptUrl = await pathProvider.GetTempUrlAsync(stream, ".docbuilder");
+                }
+            }
+            scriptUrl = ReplaceCommunityAddress(scriptUrl);
             requestKey = scriptUrl;
         }
 
@@ -143,12 +155,13 @@ public class DocumentServiceConnector(ILogger<DocumentServiceConnector> logger,
         try
         {
             return await ASC.Files.Core.Helpers.DocumentService.DocbuilderRequestAsync(
-                fileUtility,
                 filesLinkUtility.DocServiceDocbuilderUrl,
                 GenerateRevisionId(requestKey),
                 scriptUrl,
                 isAsync,
-                fileUtility.SignatureSecret,
+                filesLinkUtility.DocServiceSignatureSecret,
+                filesLinkUtility.DocServiceSignatureHeader,
+                await filesLinkUtility.GetDocServiceSslVerificationAsync(),
                 clientFactory);
         }
         catch (Exception ex)
@@ -163,14 +176,15 @@ public class DocumentServiceConnector(ILogger<DocumentServiceConnector> logger,
         try
         {
             var commandResponse = await CommandRequestAsync(
-                fileUtility,
                 filesLinkUtility.DocServiceCommandUrl,
                 CommandMethod.Version,
-                GenerateRevisionId(null),
                 null,
                 null,
                 null,
-                fileUtility.SignatureSecret,
+                null,
+                filesLinkUtility.DocServiceSignatureSecret,
+                filesLinkUtility.DocServiceSignatureHeader,
+                await filesLinkUtility.GetDocServiceSslVerificationAsync(),
                 clientFactory);
 
             var version = commandResponse.Version;
@@ -210,7 +224,10 @@ public class DocumentServiceConnector(ILogger<DocumentServiceConnector> logger,
                     Method = HttpMethod.Head
                 };
 
-                using var httpClient = clientFactory.CreateClient();
+                using var httpClient = await filesLinkUtility.GetDocServiceSslVerificationAsync()
+                    ? clientFactory.CreateClient()
+                    : clientFactory.CreateClient(CustomSslVerificationClient);
+
                 using var response = await httpClient.SendAsync(request);
 
                 if (response.StatusCode != HttpStatusCode.OK)
@@ -226,11 +243,12 @@ public class DocumentServiceConnector(ILogger<DocumentServiceConnector> logger,
             }
         }
 
-        if (!string.IsNullOrEmpty(filesLinkUtility.DocServiceHealthcheckUrl))
+        var docServiceHealthcheckUrl = filesLinkUtility.DocServiceHealthcheckUrl;
+        if (!string.IsNullOrEmpty(docServiceHealthcheckUrl))
         {
             try
             {
-                if (!await HealthcheckRequestAsync(filesLinkUtility.DocServiceHealthcheckUrl, clientFactory))
+                if (!await HealthcheckRequestAsync(docServiceHealthcheckUrl, clientFactory))
                 {
                     throw new Exception("bad status");
                 }
@@ -252,10 +270,28 @@ public class DocumentServiceConnector(ILogger<DocumentServiceConnector> logger,
                 var toExtension = fileUtility.GetInternalExtension(fileExtension);
                 var url = pathProvider.GetEmptyFileUrl(fileExtension);
 
-                var fileUri = await ReplaceCommunityAddressAsync(url);
+                var fileUri = ReplaceCommunityAddress(url);
 
                 var key = GenerateRevisionId(Guid.NewGuid().ToString());
-                var uriTuple = await ASC.Files.Core.Helpers.DocumentService.GetConvertedUriAsync(fileUtility, filesLinkUtility.DocServiceConverterUrl, fileUri, fileExtension, toExtension, key, null, null, null, null, null, false, fileUtility.SignatureSecret, clientFactory, false);
+
+                var uriTuple = await ASC.Files.Core.Helpers.DocumentService.GetConvertedUriAsync(
+                    filesLinkUtility.DocServiceConverterUrl,
+                    fileUri,
+                    fileExtension,
+                    toExtension,
+                    key,
+                    null,
+                    null,
+                    null,
+                    null,
+                    null,
+                    false,
+                    filesLinkUtility.DocServiceSignatureSecret,
+                    filesLinkUtility.DocServiceSignatureHeader,
+                    await filesLinkUtility.GetDocServiceSslVerificationAsync(),
+                    clientFactory,
+                    false);
+
                 convertedFileUri = uriTuple.ConvertedDocumentUri;
             }
             catch (Exception ex)
@@ -293,7 +329,18 @@ public class DocumentServiceConnector(ILogger<DocumentServiceConnector> logger,
             try
             {
                 var key = GenerateRevisionId(Guid.NewGuid().ToString());
-                await CommandRequestAsync(fileUtility, filesLinkUtility.DocServiceCommandUrl, CommandMethod.Version, key, null, null, null, fileUtility.SignatureSecret, clientFactory);
+
+                await CommandRequestAsync(
+                    filesLinkUtility.DocServiceCommandUrl,
+                    CommandMethod.Version,
+                    key,
+                    null,
+                    null,
+                    null,
+                    filesLinkUtility.DocServiceSignatureSecret,
+                    filesLinkUtility.DocServiceSignatureHeader,
+                    await filesLinkUtility.GetDocServiceSslVerificationAsync(),
+                    clientFactory);
             }
             catch (Exception ex)
             {
@@ -310,9 +357,17 @@ public class DocumentServiceConnector(ILogger<DocumentServiceConnector> logger,
                 var storeTemplate = await globalStore.GetStoreTemplateAsync();
                 var scriptUri = await storeTemplate.GetUriAsync("", "test.docbuilder");
                 var scriptUrl = baseCommonLinkUtility.GetFullAbsolutePath(scriptUri.ToString());
-                scriptUrl = await ReplaceCommunityAddressAsync(scriptUrl);
+                scriptUrl = ReplaceCommunityAddress(scriptUrl);
 
-                await ASC.Files.Core.Helpers.DocumentService.DocbuilderRequestAsync(fileUtility, filesLinkUtility.DocServiceDocbuilderUrl, null, scriptUrl, false, fileUtility.SignatureSecret, clientFactory);
+                await ASC.Files.Core.Helpers.DocumentService.DocbuilderRequestAsync(
+                    filesLinkUtility.DocServiceDocbuilderUrl,
+                    null,
+                    scriptUrl,
+                    false,
+                    filesLinkUtility.DocServiceSignatureSecret,
+                    filesLinkUtility.DocServiceSignatureHeader,
+                    await filesLinkUtility.GetDocServiceSslVerificationAsync(),
+                    clientFactory);
             }
             catch (Exception ex)
             {
@@ -323,7 +378,7 @@ public class DocumentServiceConnector(ILogger<DocumentServiceConnector> logger,
         }
     }
 
-    public async Task<string> ReplaceCommunityAddressAsync(string url)
+    public string ReplaceCommunityAddress(string url)
     {
         var docServicePortalUrl = filesLinkUtility.GetDocServicePortalUrl();
 
@@ -334,7 +389,7 @@ public class DocumentServiceConnector(ILogger<DocumentServiceConnector> logger,
 
         if (string.IsNullOrEmpty(docServicePortalUrl))
         {
-            var tenant = await tenantManager.GetCurrentTenantAsync();
+            var tenant = tenantManager.GetCurrentTenant();
             if (!tenantExtra.Saas
                 || string.IsNullOrEmpty(tenant.MappedDomain)
                 || !url.StartsWith("https://" + tenant.MappedDomain))
